% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert}
\alias{convert}
\title{Conversion of units}
\usage{
convert(x, from_unit, to_unit)
}
\arguments{
\item{x}{A number or numeric vector}

\item{from_unit}{Name of the unit to convert from}

\item{to_unit}{Name of the unit to convert to}
}
\value{
Converted values
}
\description{
This function lets you convert units of length, area, volume
}
\examples{
convert(1, 'Centimeter', 'Meter')
convert(100, 'Meter', 'Centimeter')
}
